simplify cli transports
