+++ crates/codestory-cli/src/main.rs
@@ -4,18 +4,16 @@ use clap_complete::{Shell, generate};
 use codestory_contracts::api::{
     AgentAnswerDto, AgentAskRequest, AgentConnectionSettingsDto, AgentHybridWeightsDto,
     AgentResponseModeDto, AppEventPayload, BookmarkCategoryDto, BookmarkDto,
-    CreateBookmarkCategoryRequest, CreateBookmarkRequest, GraphArtifactDto, GroundingBudgetDto,
-    IndexFreshnessDto, IndexFreshnessStatusDto, IndexMode, LayoutDirection,
-    ListChildrenSymbolsRequest, ListRootSymbolsRequest, NodeId, NodeKind, RepoTextScanStatsDto,
+    CreateBookmarkCategoryRequest, CreateBookmarkRequest, GraphArtifactDto, IndexFreshnessDto,
+    IndexFreshnessStatusDto, IndexMode, NodeId, NodeKind, RepoTextScanStatsDto,
     RetrievalScoreBreakdownDto, SearchHit, SearchHybridLimitsDto, SearchRepoTextMode,
-    SearchRequest, SnippetContextDto, SymbolContextDto, TrailCallerScope, TrailConfigDto,
-    TrailContextDto, TrailDirection, TrailMode,
+    SearchRequest,
 };
 use std::{
-    collections::{HashMap, HashSet},
+    collections::HashMap,
     fs,
-    io::{BufRead, IsTerminal, Read, Write},
-    net::{TcpListener, TcpStream},
+    io::IsTerminal,
+    net::TcpListener,
     sync::{
         Arc,
         atomic::{AtomicBool, Ordering},
@@ -26,38 +24,46 @@ use std::{
 mod args;
 mod config;
 mod display;
+mod explore;
+mod http_transport;
 mod output;
 mod query_resolution;
 mod runtime;
 mod stdio_catalog;
+mod stdio_transport;
 
 use args::{
     AskCommand, BookmarkAction, BookmarkAddCommand, BookmarkAddOutput, BookmarkCommand,
     BookmarkListCommand, BookmarkListOutput, BookmarkOutput, BookmarkRemoveCommand,
     BookmarkRemoveOutput, Cli, Command, CompletionShell, DoctorCheckOutput, DoctorCommand,
-    DoctorOutput, ExploreCommand, ExploreOutput, ExploreSearchOutput, ExploreStatusOutput,
-    GenerateCompletionsCommand, GroundCommand, IndexCommand, IndexDryRunOutput, IndexOutput,
-    NavigationOutput, QueryCommand, QueryItemOutput, QueryOutput, QueryResolutionOutput,
-    RepoTextMode, SearchCommand, SearchHitOutput, SearchOutput, ServeCommand, SnippetCommand,
-    SnippetJsonOutput, SymbolCommand, SymbolJsonOutput, TrailCommand, TrailJsonOutput,
-    ask_retrieval_profile, build_trail_request,
+    DoctorOutput, GenerateCompletionsCommand, GroundCommand, IndexCommand, IndexDryRunOutput,
+    IndexOutput, QueryCommand, QueryOutput, QueryResolutionOutput, RepoTextMode, SearchCommand,
+    SearchHitOutput, SearchOutput, ServeCommand, SnippetCommand, SnippetJsonOutput, SymbolCommand,
+    SymbolJsonOutput, TrailCommand, TrailJsonOutput, ask_retrieval_profile, build_trail_request,
 };
+#[cfg(test)]
+use codestory_contracts::api::TrailContextDto;
+#[cfg(test)]
+use explore::{ExploreTuiAction, ExploreTuiState, explore_tui_action};
+#[cfg(test)]
+use http_transport::search_repo_text_mode_param;
 use output::{
     emit, emit_text, render_agent_answer_markdown, render_doctor_markdown, render_ground_markdown,
     render_index_dry_run_markdown, render_index_markdown, render_query_markdown,
-    render_retrieval_state, render_search_markdown, render_snippet_markdown,
-    render_symbol_markdown, render_symbol_mermaid, render_trail_dot, render_trail_markdown,
-    render_trail_mermaid, render_trail_story_markdown, validate_output_file_parent,
+    render_search_markdown, render_snippet_markdown, render_symbol_markdown, render_symbol_mermaid,
+    render_trail_dot, render_trail_markdown, render_trail_mermaid, render_trail_story_markdown,
+    validate_output_file_parent,
 };
 use runtime::{
     AmbiguousTargetError, RuntimeContext, ensure_index_ready, map_api_error, refresh_label,
     resolve_refresh_request, resolve_target,
 };
+#[cfg(test)]
+use std::collections::HashSet;
+#[cfg(test)]
 use stdio_catalog::{
-    is_tool_name as is_stdio_tool_name, prompt_get_json as stdio_prompt_get_json,
-    prompts_list_json as stdio_prompts_list_json,
-    resource_templates_list_json as stdio_resource_templates_list_json,
-    resources_list_json as stdio_resources_list_json, tools_list_json as stdio_tools_list_json,
+    prompts_list_json as stdio_prompts_list_json, resources_list_json as stdio_resources_list_json,
+    tools_list_json as stdio_tools_list_json,
 };
 
 #[derive(Debug, Clone, Copy)]
@@ -70,14 +76,6 @@ const ASK_BUNDLE_OUTPUT_BYTE_CAP: usize = 5 * 1024 * 1024;
 const ASK_BUNDLE_MARKDOWN_SOFT_CAP: usize = 2 * 1024 * 1024;
 const ASK_BUNDLE_TRUNCATION_SUFFIX: &str =
     "\n\n... bundle content truncated by ask bundle byte cap\n";
-
-const BROWSER_TRAIL_DEFAULT_DEPTH: u32 = 2;
-const BROWSER_TRAIL_MAX_DEPTH: u32 = 10;
-const BROWSER_TRAIL_MAX_NODES: u32 = 80;
-const BROWSER_REFERENCES_DEPTH: u32 = 0;
-const BROWSER_REFERENCES_MAX_NODES: u32 = 120;
-const BROWSER_SYMBOLS_DEFAULT_LIMIT: u32 = 300;
-const BROWSER_SYMBOLS_MAX_LIMIT: u32 = 2_000;
 
 fn to_api_repo_text_mode(mode: RepoTextMode) -> SearchRepoTextMode {
     match mode {
@@ -124,7 +122,7 @@ fn main() -> Result<()> {
         Command::Trail(cmd) => run_trail(cmd),
         Command::Snippet(cmd) => run_snippet(cmd),
         Command::Query(cmd) => run_query(cmd),
-        Command::Explore(cmd) => run_explore(cmd),
+        Command::Explore(cmd) => explore::run_explore(cmd),
         Command::Bookmark(cmd) => run_bookmark(cmd),
         Command::Serve(cmd) => run_serve(cmd),
         Command::GenerateCompletions(cmd) => run_generate_completions(cmd),
@@ -271,25 +269,31 @@ fn spawn_progress_printer(rx: crossbeam_channel::Receiver<AppEventPayload>) -> P
     let handle = std::thread::spawn(move || {
         while !worker_done.load(Ordering::SeqCst) {
             match rx.recv_timeout(Duration::from_millis(100)) {
-                Ok(AppEventPayload::IndexingProgress { current, total }) => {
-                    eprintln!(
-                        "[{current}/{total}] {} indexing",
-                        format_progress_bar(current, total)
-                    );
-                }
-                Ok(AppEventPayload::IndexingStarted { file_count }) => {
-                    eprintln!(
-                        "[0/{file_count}] {} indexing started",
-                        format_progress_bar(0, file_count)
-                    );
-                }
-                Ok(_) => {}
+                Ok(event) => print_progress_event(event),
                 Err(crossbeam_channel::RecvTimeoutError::Timeout) => {}
                 Err(crossbeam_channel::RecvTimeoutError::Disconnected) => break,
             }
         }
     });
     ProgressPrinter { done, handle }
+}
+
+fn print_progress_event(event: AppEventPayload) {
+    match event {
+        AppEventPayload::IndexingProgress { current, total } => {
+            eprintln!(
+                "[{current}/{total}] {} indexing",
+                format_progress_bar(current, total)
+            );
+        }
+        AppEventPayload::IndexingStarted { file_count } => {
+            eprintln!(
+                "[0/{file_count}] {} indexing started",
+                format_progress_bar(0, file_count)
+            );
+        }
+        _ => {}
+    }
 }
 
 fn format_progress_bar(current: u32, total: u32) -> String {
@@ -661,21 +665,32 @@ fn run_search(cmd: SearchCommand) -> Result<()> {
     let runtime = RuntimeContext::new(&cmd.project)?;
     let opened = runtime.ensure_open(cmd.refresh)?;
     ensure_index_ready(&opened, "search")?;
-    let hybrid_weights = hybrid_weights(cmd.hybrid_lexical, cmd.hybrid_semantic, cmd.hybrid_graph);
-    let hybrid_limits = hybrid_limits(cmd.hybrid_lexical_limit, cmd.hybrid_semantic_limit);
-
     let search_results = runtime
         .browser
-        .search_results(SearchRequest {
-            query: cmd.query.clone(),
-            repo_text: to_api_repo_text_mode(cmd.repo_text),
-            limit_per_source: cmd.limit.clamp(1, 50),
-            hybrid_weights,
-            hybrid_limits,
-        })
+        .search_results(search_request_from_command(&cmd))
         .map_err(map_api_error)?;
-    let output = build_search_output(
-        &runtime.project_root,
+    let output = search_output_from_results(&runtime.project_root, &search_results, cmd.why);
+    let markdown = render_search_markdown(&runtime.project_root, &output);
+    emit(cmd.format, &output, markdown, cmd.output_file.as_deref())
+}
+
+fn search_request_from_command(cmd: &SearchCommand) -> SearchRequest {
+    SearchRequest {
+        query: cmd.query.clone(),
+        repo_text: to_api_repo_text_mode(cmd.repo_text),
+        limit_per_source: cmd.limit.clamp(1, 50),
+        hybrid_weights: hybrid_weights(cmd.hybrid_lexical, cmd.hybrid_semantic, cmd.hybrid_graph),
+        hybrid_limits: hybrid_limits(cmd.hybrid_lexical_limit, cmd.hybrid_semantic_limit),
+    }
+}
+
+fn search_output_from_results(
+    project_root: &std::path::Path,
+    search_results: &codestory_contracts::api::SearchResultsDto,
+    include_score_details: bool,
+) -> SearchOutput {
+    build_search_output(
+        project_root,
         &search_results.query,
         &search_results.retrieval,
         search_results.freshness.as_ref(),
@@ -688,10 +703,8 @@ fn run_search(cmd: SearchCommand) -> Result<()> {
             mode: from_api_repo_text_mode(search_results.repo_text_mode),
             enabled: search_results.repo_text_enabled,
         },
-        cmd.why,
-    );
-    let markdown = render_search_markdown(&runtime.project_root, &output);
-    emit(cmd.format, &output, markdown, cmd.output_file.as_deref())
+        include_score_details,
+    )
 }
 
 fn run_symbol(cmd: SymbolCommand) -> Result<()> {
@@ -813,7 +826,7 @@ fn run_query(cmd: QueryCommand) -> Result<()> {
         .query(&ast)
         .map_err(map_api_error)?
         .iter()
-        .map(|item| browser_query_item_to_output(&runtime.project_root, item))
+        .map(|item| explore::browser_query_item_to_output(&runtime.project_root, item))
         .collect();
     let output = QueryOutput {
         query: cmd.query,
@@ -824,111 +837,12 @@ fn run_query(cmd: QueryCommand) -> Result<()> {
     emit(cmd.format, &output, markdown, cmd.output_file.as_deref())
 }
 
-fn run_explore(cmd: ExploreCommand) -> Result<()> {
-    ensure_dot_only_for_trail(cmd.format, "explore")?;
-    preflight_output_file(cmd.output_file.as_deref())?;
-    let runtime = RuntimeContext::new(&cmd.project)?;
-    let opened = runtime.ensure_open(cmd.refresh)?;
-    ensure_index_ready(&opened, "explore")?;
-    let file_filter = cmd.target.file_filter();
-    let target = resolve_target_or_emit_ambiguity(
-        &runtime,
-        cmd.target.selection()?,
-        file_filter.as_deref(),
-        cmd.format,
-        cmd.output_file.as_deref(),
-    )?;
-    let symbol = runtime
-        .browser
-        .symbol_context(target.selected.node_id.clone())
-        .map_err(map_api_error)?;
-    let trail = runtime
-        .browser
-        .trail_context(TrailConfigDto {
-            root_id: target.selected.node_id.clone(),
-            mode: TrailMode::Neighborhood,
-            target_id: None,
-            depth: cmd.depth,
-            direction: TrailDirection::Both,
-            caller_scope: TrailCallerScope::ProductionOnly,
-            edge_filter: Vec::new(),
-            show_utility_calls: false,
-            hide_speculative: false,
-            story: false,
-            node_filter: Vec::new(),
-            max_nodes: cmd.max_nodes.clamp(1, 120),
-            layout_direction: LayoutDirection::Horizontal,
-        })
-        .map_err(map_api_error)?;
-    let snippet_result = runtime
-        .browser
-        .snippet_context(target.selected.node_id.clone(), 4);
-    let (snippet, snippet_layer_note) = match snippet_result {
-        Ok(snippet) => (Some(snippet), "snippet_context: available".to_string()),
-        Err(error) => (
-            None,
-            format!(
-                "snippet_context: unavailable: {}: {}",
-                error.code, error.message
-            ),
-        ),
-    };
-    let status = build_explore_status_output(
-        &runtime,
-        &opened,
-        &target,
-        cmd.refresh,
-        cmd.output_file.as_deref(),
-        &snippet_layer_note,
-    );
-    let search = build_explore_search_output(&runtime.project_root, &target);
-    let navigation = build_navigation_output(&runtime.project_root, &target, &trail);
-    let output = ExploreOutput {
-        status,
-        search,
-        resolution: build_query_resolution_output(&runtime.project_root, &target),
-        navigation,
-        symbol: &symbol,
-        trail: &trail,
-        snippet: snippet.as_ref(),
-    };
-    let markdown = render_explore_markdown(
-        &runtime.project_root,
-        &target,
-        &output.status,
-        &output.search,
-        &output.navigation,
-        &symbol,
-        &trail,
-        snippet.as_ref(),
-        &snippet_layer_note,
-    );
-    if cmd.format == args::OutputFormat::Markdown
-        && cmd.output_file.is_none()
-        && !cmd.no_tui
-        && std::io::stdout().is_terminal()
-    {
-        return run_explore_tui(
-            &runtime.project_root,
-            &target,
-            &output.status,
-            &output.search,
-            &output.navigation,
-            &symbol,
-            &trail,
-            snippet.as_ref(),
-            &snippet_layer_note,
-        );
-    }
-    emit(cmd.format, &output, markdown, cmd.output_file.as_deref())
-}
-
 fn run_serve(cmd: ServeCommand) -> Result<()> {
     let runtime = RuntimeContext::new(&cmd.project)?;
     let opened = runtime.ensure_open(cmd.refresh)?;
     ensure_index_ready(&opened, "serve")?;
     if cmd.stdio {
-        return run_stdio_server(runtime);
+        return stdio_transport::run_stdio_server(runtime);
     }
     let listener = TcpListener::bind(&cmd.addr)
         .with_context(|| format!("Failed to bind server to {}", cmd.addr))?;
@@ -936,7 +850,7 @@ fn run_serve(cmd: ServeCommand) -> Result<()> {
     for stream in listener.incoming() {
         match stream {
             Ok(stream) => {
-                if let Err(error) = handle_http_request(&runtime, stream) {
+                if let Err(error) = http_transport::handle_http_request(&runtime, stream) {
                     eprintln!("serve request failed: {error:#}");
                 }
             }
@@ -1542,1880 +1456,6 @@ fn sanitize_artifact_name(value: &str) -> String {
     out
 }
 
-fn graph_node_to_query_item(
-    project_root: &std::path::Path,
-    node: &codestory_contracts::api::GraphNodeDto,
-    source: &str,
-) -> QueryItemOutput {
-    let file_path = node
-        .file_path
-        .as_deref()
-        .map(|path| display::relative_path(project_root, path));
-    QueryItemOutput {
-        node_id: node.id.0.clone(),
-        node_ref: None,
-        display_name: node.label.clone(),
-        kind: node.kind,
-        file_path,
-        line: None,
-        depth: Some(node.depth),
-        source: source.to_string(),
-    }
-}
-
-fn browser_query_item_to_output(
-    project_root: &std::path::Path,
-    item: &codestory_runtime::BrowserQueryItem,
-) -> QueryItemOutput {
-    QueryItemOutput {
-        node_id: item.node_id.0.clone(),
-        node_ref: output::node_ref(
-            project_root,
-            item.file_path.as_deref(),
-            item.line,
-            &item.display_name,
-        ),
-        display_name: item.display_name.clone(),
-        kind: item.kind,
-        file_path: item
-            .file_path
-            .as_deref()
-            .map(|path| display::relative_path(project_root, path)),
-        line: item.line,
-        depth: item.depth,
-        source: item.source.clone(),
-    }
-}
-
-fn build_navigation_output(
-    project_root: &std::path::Path,
-    target: &runtime::ResolvedTarget,
-    trail: &TrailContextDto,
-) -> NavigationOutput {
-    let center = &target.selected.node_id;
-    let nodes = trail
-        .trail
-        .nodes
-        .iter()
-        .map(|node| (node.id.clone(), node))
-        .collect::<HashMap<_, _>>();
-    let mut incoming_seen = HashSet::new();
-    let mut outgoing_seen = HashSet::new();
-    let mut incoming_references = Vec::new();
-    let mut outgoing_references = Vec::new();
-
-    for edge in &trail.trail.edges {
-        if &edge.target == center
-            && incoming_seen.insert(edge.source.clone())
-            && let Some(node) = nodes.get(&edge.source)
-        {
-            incoming_references.push(graph_node_to_query_item(
-                project_root,
-                node,
-                "incoming_reference",
-            ));
-        }
-        if &edge.source == center
-            && outgoing_seen.insert(edge.target.clone())
-            && let Some(node) = nodes.get(&edge.target)
-        {
-            outgoing_references.push(graph_node_to_query_item(
-                project_root,
-                node,
-                "outgoing_reference",
-            ));
-        }
-    }
-
-    NavigationOutput {
-        definition: build_search_hit_output(project_root, &target.selected, false),
-        incoming_references,
-        outgoing_references,
-    }
-}
-
-fn build_explore_search_output(
-    project_root: &std::path::Path,
-    target: &runtime::ResolvedTarget,
-) -> ExploreSearchOutput {
-    ExploreSearchOutput {
-        selector: target.selector,
-        requested: target.requested.clone(),
-        file_filter: target
-            .file_filter
-            .as_deref()
-            .map(crate::display::clean_path_string),
-        selected: build_search_hit_output(project_root, &target.selected, false),
-        alternatives: target
-            .alternatives
-            .iter()
-            .skip(1)
-            .map(|hit| build_search_hit_output(project_root, hit, false))
-            .collect(),
-    }
-}
-
-fn build_explore_status_output(
-    runtime: &RuntimeContext,
-    opened: &runtime::OpenedProject,
-    target: &runtime::ResolvedTarget,
-    requested_refresh: args::RefreshMode,
-    output_file: Option<&std::path::Path>,
-    snippet_layer_note: &str,
-) -> ExploreStatusOutput {
-    let project = display::clean_path_string(&runtime.project_root.to_string_lossy());
-    let storage_path = display::clean_path_string(&runtime.storage_path.to_string_lossy());
-    let output_target = output_file
-        .map(|path| display::clean_path_string(&path.to_string_lossy()))
-        .unwrap_or_else(|| "stdout".to_string());
-    let retrieval = opened.summary.retrieval.clone();
-    let freshness = opened.summary.freshness.clone();
-    let next_commands =
-        explore_next_commands(&project, target, retrieval.as_ref(), freshness.as_ref());
-    let layer_notes = explore_layer_notes(
-        &storage_path,
-        &output_target,
-        &opened.summary,
-        target,
-        retrieval.as_ref(),
-        freshness.as_ref(),
-        snippet_layer_note,
-    );
-
-    ExploreStatusOutput {
-        project,
-        storage_path,
-        refresh: refresh_label(requested_refresh, opened.refresh_mode),
-        output_target,
-        indexed_files: opened.summary.stats.file_count,
-        indexed_nodes: opened.summary.stats.node_count,
-        indexed_edges: opened.summary.stats.edge_count,
-        retrieval,
-        freshness,
-        next_commands,
-        layer_notes,
-    }
-}
-
-fn explore_next_commands(
-    project: &str,
-    target: &runtime::ResolvedTarget,
-    retrieval: Option<&codestory_contracts::api::RetrievalStateDto>,
-    freshness: Option<&IndexFreshnessDto>,
-) -> Vec<String> {
-    let node_id = &target.selected.node_id.0;
-    let mut commands = Vec::new();
-    if freshness.is_some_and(|state| state.status == IndexFreshnessStatusDto::Stale) {
-        commands.push(format!(
-            "codestory-cli index --project \"{project}\" --refresh incremental"
-        ));
-    }
-    commands.push(format!(
-        "codestory-cli ask --project \"{project}\" --focus-id {node_id} \"What should I inspect next?\""
-    ));
-    commands.push(format!(
-        "codestory-cli trail --project \"{project}\" --id {node_id} --depth 3"
-    ));
-    commands.push(format!(
-        "codestory-cli snippet --project \"{project}\" --id {node_id}"
-    ));
-    if retrieval.is_some_and(|state| !state.semantic_ready) {
-        commands.push(format!(
-            "codestory-cli doctor --project \"{project}\" --format markdown"
-        ));
-    }
-    commands
-}
-
-fn explore_layer_notes(
-    storage_path: &str,
-    output_target: &str,
-    summary: &codestory_contracts::api::ProjectSummary,
-    target: &runtime::ResolvedTarget,
-    retrieval: Option<&codestory_contracts::api::RetrievalStateDto>,
-    freshness: Option<&IndexFreshnessDto>,
-    snippet_layer_note: &str,
-) -> Vec<String> {
-    let mut notes = vec![
-        format!("cache: reading existing SQLite cache at `{storage_path}`"),
-        format!(
-            "index: ready with files={} nodes={} edges={}",
-            summary.stats.file_count, summary.stats.node_count, summary.stats.edge_count
-        ),
-        format!(
-            "query_resolution: `{}` resolved to `{}`",
-            target.requested, target.selected.display_name
-        ),
-        "local_agent: not used by explore; use ask --focus-id for agent synthesis".to_string(),
-        format!("output_write: target `{output_target}` passed preflight"),
-        snippet_layer_note.to_string(),
-    ];
-    notes.push(match retrieval {
-        Some(retrieval) if retrieval.semantic_ready => {
-            format!(
-                "semantic_runtime: ready with {} semantic docs",
-                retrieval.semantic_doc_count
-            )
-        }
-        Some(retrieval) => format!(
-            "semantic_runtime: {}",
-            retrieval
-                .fallback_message
-                .as_deref()
-                .unwrap_or("semantic retrieval fallback is active")
-        ),
-        None => "semantic_runtime: retrieval state unavailable".to_string(),
-    });
-    notes.push(match freshness {
-        Some(freshness) => format!("freshness: {}", render_explore_freshness(freshness)),
-        None => "freshness: not reported by this cache open".to_string(),
-    });
-    notes
-}
-
-fn render_explore_freshness(freshness: &IndexFreshnessDto) -> String {
-    match freshness.status {
-        IndexFreshnessStatusDto::Fresh => format!(
-            "fresh checked_files={} duration_ms={}",
-            freshness.checked_file_count, freshness.duration_ms
-        ),
-        IndexFreshnessStatusDto::Stale => format!(
-            "stale changed={} new={} removed={} checked_files={} duration_ms={}",
-            freshness.changed_file_count,
-            freshness.new_file_count,
-            freshness.removed_file_count,
-            freshness.checked_file_count,
-            freshness.duration_ms
-        ),
-        IndexFreshnessStatusDto::NotChecked => format!(
-            "not_checked reason={}",
-            freshness.reason.as_deref().unwrap_or("not reported")
-        ),
-    }
-}
-
-fn render_explore_status_markdown(status: &ExploreStatusOutput) -> String {
-    let mut markdown = String::new();
-    markdown.push_str(&format!("- project: `{}`\n", status.project));
-    markdown.push_str(&format!("- cache: `{}`\n", status.storage_path));
-    markdown.push_str(&format!("- refresh: `{}`\n", status.refresh));
-    markdown.push_str(&format!("- output: `{}`\n", status.output_target));
-    markdown.push_str(&format!(
-        "- indexed: files={} nodes={} edges={}\n",
-        status.indexed_files, status.indexed_nodes, status.indexed_edges
-    ));
-    if let Some(retrieval) = status.retrieval.as_ref() {
-        markdown.push_str(&format!(
-            "- retrieval: {}\n",
-            render_retrieval_state(retrieval)
-        ));
-    } else {
-        markdown.push_str("- retrieval: unavailable\n");
-    }
-    if let Some(freshness) = status.freshness.as_ref() {
-        markdown.push_str(&format!(
-            "- freshness: {}\n",
-            render_explore_freshness(freshness)
-        ));
-    } else {
-        markdown.push_str("- freshness: unavailable\n");
-    }
-    if let Some(command) = status.next_commands.first() {
-        markdown.push_str(&format!("- next: `{command}`\n"));
-    }
-    markdown.push_str("- layers:\n");
-    for note in &status.layer_notes {
-        markdown.push_str(&format!("  - {note}\n"));
-    }
-    markdown
-}
-
-fn format_query_selector(selector: args::QuerySelectorOutput) -> &'static str {
-    match selector {
-        args::QuerySelectorOutput::Id => "id",
-        args::QuerySelectorOutput::Query => "query",
-    }
-}
-
-fn render_search_hit_output_ref(hit: &SearchHitOutput) -> String {
-    if let Some(node_ref) = hit.node_ref.as_deref() {
-        return format!("`{node_ref}`");
-    }
-    if let Some(file_path) = hit.file_path.as_deref() {
-        if let Some(line) = hit.line {
-            return format!("{} `{file_path}:{line}`", hit.display_name);
-        }
-        return format!("{} `{file_path}`", hit.display_name);
-    }
-    hit.display_name.clone()
-}
-
-fn render_query_item_output_ref(item: &QueryItemOutput) -> String {
-    if let Some(node_ref) = item.node_ref.as_deref() {
-        return format!("`{node_ref}`");
-    }
-    if let Some(file_path) = item.file_path.as_deref() {
-        if let Some(line) = item.line {
-            return format!("{} `{file_path}:{line}`", item.display_name);
-        }
-        return format!("{} `{file_path}`", item.display_name);
-    }
-    item.display_name.clone()
-}
-
-fn render_explore_search_markdown(search: &ExploreSearchOutput) -> String {
-    let mut markdown = String::new();
-    markdown.push_str(&format!(
-        "- selector: `{}`\n",
-        format_query_selector(search.selector)
-    ));
-    markdown.push_str(&format!("- requested: `{}`\n", search.requested));
-    if let Some(file_filter) = search.file_filter.as_deref() {
-        markdown.push_str(&format!("- file_filter: `{file_filter}`\n"));
-    }
-    markdown.push_str(&format!(
-        "- selected: {}\n",
-        render_search_hit_output_ref(&search.selected)
-    ));
-    markdown.push_str(&format!("- alternatives: {}\n", search.alternatives.len()));
-    for (index, alternative) in search.alternatives.iter().take(8).enumerate() {
-        markdown.push_str(&format!(
-            "  - {}. {}\n",
-            index + 1,
-            render_search_hit_output_ref(alternative)
-        ));
-    }
-    markdown
-}
-
-fn render_explore_results_markdown(navigation: &NavigationOutput) -> String {
-    let mut markdown = String::new();
-    markdown.push_str(&format!(
-        "- definition: {}\n",
-        render_search_hit_output_ref(&navigation.definition)
-    ));
-    markdown.push_str(&format!(
-        "- incoming_references: {}\n",
-        navigation.incoming_references.len()
-    ));
-    markdown.push_str(&format!(
-        "- outgoing_references: {}\n",
-        navigation.outgoing_references.len()
-    ));
-    for incoming in navigation.incoming_references.iter().take(6) {
-        markdown.push_str(&format!(
-            "- incoming: {}\n",
-            render_query_item_output_ref(incoming)
-        ));
-    }
-    for outgoing in navigation.outgoing_references.iter().take(6) {
-        markdown.push_str(&format!(
-            "- outgoing: {}\n",
-            render_query_item_output_ref(outgoing)
-        ));
-    }
-    markdown
-}
-
-fn render_explore_markdown(
-    project_root: &std::path::Path,
-    target: &runtime::ResolvedTarget,
-    status: &ExploreStatusOutput,
-    search: &ExploreSearchOutput,
-    navigation: &NavigationOutput,
-    symbol: &SymbolContextDto,
-    trail: &TrailContextDto,
-    snippet: Option<&SnippetContextDto>,
-    snippet_layer_note: &str,
-) -> String {
-    let mut markdown = String::new();
-    markdown.push_str("# Explore\n");
-    markdown.push_str("status:\n");
-    markdown.push_str(&render_explore_status_markdown(status));
-    markdown.push_str("search:\n");
-    markdown.push_str(&render_explore_search_markdown(search));
-    markdown.push_str("results:\n");
-    markdown.push_str(&render_explore_results_markdown(navigation));
-    markdown.push_str("resolution:\n");
-    markdown.push_str(&format!(
-        "- {}\n",
-        output::node_ref(
-            project_root,
-            target.selected.file_path.as_deref(),
-            target.selected.line,
-            &target.selected.display_name
-        )
-        .unwrap_or_else(|| target.selected.display_name.clone())
-    ));
-    markdown.push_str("navigation:\n");
-    if let Some(node_ref) = navigation.definition.node_ref.as_deref() {
-        markdown.push_str(&format!("- definition: `{node_ref}`\n"));
-    } else {
-        markdown.push_str(&format!(
-            "- definition: {}\n",
-            navigation.definition.display_name
-        ));
-    }
-    markdown.push_str(&format!(
-        "- incoming_references: {}\n",
-        navigation.incoming_references.len()
-    ));
-    markdown.push_str(&format!(
-        "- outgoing_references: {}\n",
-        navigation.outgoing_references.len()
-    ));
-    markdown.push_str("symbol:\n");
-    markdown.push_str(&render_symbol_markdown(project_root, target, symbol));
-    markdown.push_str("\ntrail:\n");
-    let cmd = TrailCommand {
-        project: args::ProjectArgs {
-            project: project_root.to_path_buf(),
-            cache_dir: None,
-        },
-        target: args::TargetArgs {
-            id: Some(target.selected.node_id.0.clone()),
-            query: None,
-            file: None,
-            choose: None,
-        },
-        mode: args::CliTrailMode::Neighborhood,
-        depth: Some(2),
-        direction: Some(args::CliDirection::Both),
-        max_nodes: trail.trail.nodes.len().min(u32::MAX as usize) as u32,
-        include_tests: false,
-        show_utility_calls: false,
-        hide_speculative: false,
-        story: false,
-        layout: args::CliLayout::Horizontal,
-        refresh: args::RefreshMode::None,
-        format: args::OutputFormat::Markdown,
-        output_file: None,
-        mermaid: false,
-    };
-    markdown.push_str(&render_trail_markdown(project_root, target, trail, &cmd));
-    markdown.push_str("\nsnippet:\n");
-    markdown.push_str(&format!("- {snippet_layer_note}\n"));
-    if let Some(snippet) = snippet {
-        markdown.push_str(&render_snippet_markdown(
-            project_root,
-            target,
-            snippet,
-            false,
-        ));
-    }
-    markdown
-}
-
-struct ExplorePane {
-    label: &'static str,
-    body: String,
-}
-
-fn build_explore_panes(
-    project_root: &std::path::Path,
-    target: &runtime::ResolvedTarget,
-    status: &ExploreStatusOutput,
-    search: &ExploreSearchOutput,
-    navigation: &NavigationOutput,
-    symbol: &SymbolContextDto,
-    trail: &TrailContextDto,
-    snippet: Option<&SnippetContextDto>,
-    snippet_layer_note: &str,
-) -> Vec<ExplorePane> {
-    vec![
-        ExplorePane {
-            label: "Status",
-            body: render_explore_status_markdown(status),
-        },
-        ExplorePane {
-            label: "Search",
-            body: render_explore_search_markdown(search),
-        },
-        ExplorePane {
-            label: "Results",
-            body: render_explore_results_markdown(navigation),
-        },
-        ExplorePane {
-            label: "Detail",
-            body: render_symbol_markdown(project_root, target, symbol),
-        },
-        ExplorePane {
-            label: "Trail",
-            body: {
-                let cmd = TrailCommand {
-                    project: args::ProjectArgs {
-                        project: project_root.to_path_buf(),
-                        cache_dir: None,
-                    },
-                    target: args::TargetArgs {
-                        id: Some(target.selected.node_id.0.clone()),
-                        query: None,
-                        file: None,
-                        choose: None,
-                    },
-                    mode: args::CliTrailMode::Neighborhood,
-                    depth: Some(2),
-                    direction: Some(args::CliDirection::Both),
-                    max_nodes: trail.trail.nodes.len().min(u32::MAX as usize) as u32,
-                    include_tests: false,
-                    show_utility_calls: false,
-                    hide_speculative: false,
-                    story: false,
-                    layout: args::CliLayout::Horizontal,
-                    refresh: args::RefreshMode::None,
-                    format: args::OutputFormat::Markdown,
-                    output_file: None,
-                    mermaid: false,
-                };
-                render_trail_markdown(project_root, target, trail, &cmd)
-            },
-        },
-        ExplorePane {
-            label: "Snippet",
-            body: format!(
-                "{snippet_layer_note}\n{}",
-                snippet
-                    .map(|context| render_snippet_markdown(project_root, target, context, false))
-                    .unwrap_or_default()
-            ),
-        },
-    ]
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ExploreTuiAction {
-    NextPane,
-    PreviousPane,
-    ScrollUp(u16),
-    ScrollDown(u16),
-    Home,
-    Quit,
-    None,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ExploreTuiState {
-    selected: usize,
-    scroll: Vec<u16>,
-}
-
-impl ExploreTuiState {
-    fn new(pane_count: usize) -> Self {
-        Self {
-            selected: 0,
-            scroll: vec![0; pane_count.max(1)],
-        }
-    }
-
-    fn apply(&mut self, action: ExploreTuiAction) -> bool {
-        match action {
-            ExploreTuiAction::NextPane => self.selected = (self.selected + 1) % self.scroll.len(),
-            ExploreTuiAction::PreviousPane => {
-                self.selected = (self.selected + self.scroll.len() - 1) % self.scroll.len();
-            }
-            ExploreTuiAction::ScrollUp(lines) => {
-                self.scroll[self.selected] = self.scroll[self.selected].saturating_sub(lines);
-            }
-            ExploreTuiAction::ScrollDown(lines) => {
-                self.scroll[self.selected] = self.scroll[self.selected].saturating_add(lines);
-            }
-            ExploreTuiAction::Home => self.scroll[self.selected] = 0,
-            ExploreTuiAction::Quit => return true,
-            ExploreTuiAction::None => {}
-        }
-        false
-    }
-}
-
-fn explore_tui_action(key: crossterm::event::KeyEvent) -> ExploreTuiAction {
-    use crossterm::event::{KeyCode, KeyModifiers};
-    match key.code {
-        KeyCode::Esc | KeyCode::Char('q') => ExploreTuiAction::Quit,
-        KeyCode::Char('c') if key.modifiers.contains(KeyModifiers::CONTROL) => {
-            ExploreTuiAction::Quit
-        }
-        KeyCode::Tab => ExploreTuiAction::NextPane,
-        KeyCode::BackTab => ExploreTuiAction::PreviousPane,
-        KeyCode::Up | KeyCode::Char('k') => ExploreTuiAction::ScrollUp(1),
-        KeyCode::Down | KeyCode::Char('j') => ExploreTuiAction::ScrollDown(1),
-        KeyCode::PageUp => ExploreTuiAction::ScrollUp(10),
-        KeyCode::PageDown => ExploreTuiAction::ScrollDown(10),
-        KeyCode::Home => ExploreTuiAction::Home,
-        _ => ExploreTuiAction::None,
-    }
-}
-
-struct TerminalCleanup;
-
-impl Drop for TerminalCleanup {
-    fn drop(&mut self) {
-        let _ = crossterm::terminal::disable_raw_mode();
-        let _ = crossterm::execute!(std::io::stdout(), crossterm::terminal::LeaveAlternateScreen);
-    }
-}
-
-fn run_explore_tui(
-    project_root: &std::path::Path,
-    target: &runtime::ResolvedTarget,
-    status: &ExploreStatusOutput,
-    search: &ExploreSearchOutput,
-    navigation: &NavigationOutput,
-    symbol: &SymbolContextDto,
-    trail: &TrailContextDto,
-    snippet: Option<&SnippetContextDto>,
-    snippet_layer_note: &str,
-) -> Result<()> {
-    use crossterm::{
-        event::{self, Event},
-        terminal::{EnterAlternateScreen, enable_raw_mode},
-    };
-    use ratatui::{
-        Terminal,
-        backend::CrosstermBackend,
-        layout::{Constraint, Direction, Layout},
-        style::{Color, Modifier, Style},
-        text::{Line, Span},
-        widgets::{Block, Borders, List, ListItem, Paragraph, Wrap},
-    };
-
-    let panes = build_explore_panes(
-        project_root,
-        target,
-        status,
-        search,
-        navigation,
-        symbol,
-        trail,
-        snippet,
-        snippet_layer_note,
-    );
-
-    enable_raw_mode()?;
-    crossterm::execute!(std::io::stdout(), EnterAlternateScreen)?;
-    let _cleanup = TerminalCleanup;
-    let backend = CrosstermBackend::new(std::io::stdout());
-    let mut terminal = Terminal::new(backend)?;
-    let mut state = ExploreTuiState::new(panes.len());
-
-    loop {
-        terminal.draw(|frame| {
-            let area = frame.area();
-            let shell = Layout::default()
-                .direction(Direction::Vertical)
-                .constraints([
-                    Constraint::Length(3),
-                    Constraint::Min(1),
-                    Constraint::Length(1),
-                ])
-                .split(area);
-            let body = Layout::default()
-                .direction(Direction::Horizontal)
-                .constraints([Constraint::Length(24), Constraint::Min(30)])
-                .split(shell[1]);
-
-            let title = output::node_ref(
-                project_root,
-                target.selected.file_path.as_deref(),
-                target.selected.line,
-                &target.selected.display_name,
-            )
-            .unwrap_or_else(|| target.selected.display_name.clone());
-            frame.render_widget(
-                Paragraph::new(title).block(
-                    Block::default()
-                        .borders(Borders::ALL)
-                        .title("CodeStory Explore"),
-                ),
-                shell[0],
-            );
-
-            let nav_items = panes
-                .iter()
-                .enumerate()
-                .map(|(idx, pane)| {
-                    let style = if idx == state.selected {
-                        Style::default()
-                            .fg(Color::Cyan)
-                            .add_modifier(Modifier::BOLD)
-                    } else {
-                        Style::default()
-                    };
-                    ListItem::new(Line::from(Span::styled(pane.label, style)))
-                })
-                .collect::<Vec<_>>();
-            frame.render_widget(
-                List::new(nav_items).block(Block::default().borders(Borders::ALL).title("Panes")),
-                body[0],
-            );
-
-            frame.render_widget(
-                Paragraph::new(panes[state.selected].body.as_str())
-                    .block(
-                        Block::default()
-                            .borders(Borders::ALL)
-                            .title(panes[state.selected].label),
-                    )
-                    .wrap(Wrap { trim: false })
-                    .scroll((state.scroll[state.selected], 0)),
-                body[1],
-            );
-            frame.render_widget(
-                Paragraph::new("Tab/Shift-Tab pane  Up/Down scroll  Home top  q quit"),
-                shell[2],
-            );
-        })?;
-
-        if event::poll(Duration::from_millis(250))?
-            && let Event::Key(key) = event::read()?
-            && state.apply(explore_tui_action(key))
-        {
-            break;
-        }
-    }
-    terminal.show_cursor()?;
-    Ok(())
-}
-
-fn handle_http_request(runtime: &RuntimeContext, mut stream: TcpStream) -> Result<()> {
-    stream.set_read_timeout(Some(Duration::from_secs(2)))?;
-    let mut request_bytes = Vec::with_capacity(1024);
-    let mut buffer = [0u8; 1024];
-    let mut headers_complete = false;
-    loop {
-        let read = match stream.read(&mut buffer) {
-            Ok(read) => read,
-            Err(error)
-                if matches!(
-                    error.kind(),
-                    std::io::ErrorKind::TimedOut | std::io::ErrorKind::WouldBlock
-                ) =>
-            {
-                break;
-            }
-            Err(error) => return Err(error.into()),
-        };
-        if read == 0 {
-            break;
-        }
-        request_bytes.extend_from_slice(&buffer[..read]);
-        if request_bytes.windows(4).any(|window| window == b"\r\n\r\n") {
-            headers_complete = true;
-            break;
-        }
-        if request_bytes.len() >= 8192 {
-            break;
-        }
-    }
-    if !headers_complete {
-        return write_http_json(
-            &mut stream,
-            400,
-            &serde_json::json!({"error": "bad request"}),
-        );
-    }
-    let request = String::from_utf8_lossy(&request_bytes);
-    let line = request.lines().next().unwrap_or_default();
-    let mut parts = line.split_whitespace();
-    let method = parts.next().unwrap_or_default();
-    let target = parts.next().unwrap_or("/");
-    if method != "GET" {
-        return write_http_json(
-            &mut stream,
-            405,
-            &serde_json::json!({"error": "method not allowed"}),
-        );
-    }
-    let (path, query) = target.split_once('?').unwrap_or((target, ""));
-    let params = parse_query_string(query);
-    match path {
-        "/health" => write_http_json(&mut stream, 200, &serde_json::json!({"ok": true})),
-        "/search" => {
-            let query = params.get("q").cloned().unwrap_or_default();
-            let repo_text = params
-                .get("repo_text")
-                .and_then(|value| search_repo_text_mode_param(value))
-                .unwrap_or(SearchRepoTextMode::Auto);
-            let limit_per_source = params
-                .get("limit")
-                .and_then(|value| value.parse::<u32>().ok())
-                .unwrap_or(10)
-                .clamp(1, 100);
-            let results = runtime
-                .browser
-                .search_results(SearchRequest {
-                    query,
-                    repo_text,
-                    limit_per_source,
-                    hybrid_weights: None,
-                    hybrid_limits: None,
-                })
-                .map_err(map_api_error)?;
-            write_http_json(&mut stream, 200, &results)
-        }
-        "/symbol" => {
-            let Some(selection) = http_target_selection_or_error(&mut stream, &params)? else {
-                return Ok(());
-            };
-            let Some(target) = resolve_http_target(&mut stream, runtime, selection, None)? else {
-                return Ok(());
-            };
-            let context = runtime
-                .browser
-                .symbol_context(target.selected.node_id)
-                .map_err(map_api_error)?;
-            write_http_json(&mut stream, 200, &context)
-        }
-        "/definition" => {
-            let Some(selection) = http_target_selection_or_error(&mut stream, &params)? else {
-                return Ok(());
-            };
-            let Some(target) = resolve_http_target(&mut stream, runtime, selection, None)? else {
-                return Ok(());
-            };
-            let context = runtime
-                .browser
-                .definition_context(target.selected.node_id.clone())
-                .map_err(map_api_error)?;
-            write_http_json(
-                &mut stream,
-                200,
-                &serde_json::json!({
-                    "resolution": build_query_resolution_output(&runtime.project_root, &target),
-                    "definition": build_search_hit_output(&runtime.project_root, &target.selected, false),
-                    "symbol": context,
-                }),
-            )
-        }
-        "/references" => {
-            let Some(selection) = http_target_selection_or_error(&mut stream, &params)? else {
-                return Ok(());
-            };
-            let Some(target) = resolve_http_target(&mut stream, runtime, selection, None)? else {
-                return Ok(());
-            };
-            let context = runtime
-                .browser
-                .references_context(browser_references_config(target.selected.node_id.clone()))
-                .map_err(map_api_error)?;
-            write_http_json(
-                &mut stream,
-                200,
-                &serde_json::json!({
-                    "resolution": build_query_resolution_output(&runtime.project_root, &target),
-                    "references": context,
-                }),
-            )
-        }
-        "/symbols" => {
-            let limit = browser_symbols_limit(params.get("limit").map(String::as_str));
-            if let Some(parent_id) = params.get("parent_id").filter(|value| !value.is_empty()) {
-                let symbols = runtime
-                    .browser
-                    .list_children_symbols(ListChildrenSymbolsRequest {
-                        parent_id: NodeId(parent_id.clone()),
-                    })
-                    .map_err(map_api_error)?;
-                write_http_json(&mut stream, 200, &symbols)
-            } else {
-                let symbols = runtime
-                    .browser
-                    .list_root_symbols(ListRootSymbolsRequest { limit })
-                    .map_err(map_api_error)?;
-                write_http_json(&mut stream, 200, &symbols)
-            }
-        }
-        "/trail" => {
-            let Some(selection) = http_target_selection_or_error(&mut stream, &params)? else {
-                return Ok(());
-            };
-            let Some(target) = resolve_http_target(&mut stream, runtime, selection, None)? else {
-                return Ok(());
-            };
-            let depth = browser_trail_depth(params.get("depth").map(String::as_str));
-            let direction = browser_trail_direction(params.get("direction").map(String::as_str));
-            let story = browser_bool_param(params.get("story").map(String::as_str));
-            let context = runtime
-                .browser
-                .trail_context(browser_trail_config(
-                    target.selected.node_id,
-                    depth,
-                    direction,
-                    story,
-                ))
-                .map_err(map_api_error)?;
-            write_http_json(&mut stream, 200, &context)
-        }
-        _ => write_http_json(&mut stream, 404, &serde_json::json!({"error": "not found"})),
-    }
-}
-
-fn browser_references_config(root_id: NodeId) -> TrailConfigDto {
-    TrailConfigDto {
-        root_id,
-        mode: TrailMode::AllReferencing,
-        target_id: None,
-        depth: BROWSER_REFERENCES_DEPTH,
-        direction: TrailDirection::Incoming,
-        caller_scope: TrailCallerScope::IncludeTestsAndBenches,
-        edge_filter: Vec::new(),
-        show_utility_calls: false,
-        hide_speculative: false,
-        story: false,
-        node_filter: Vec::new(),
-        max_nodes: BROWSER_REFERENCES_MAX_NODES,
-        layout_direction: LayoutDirection::Horizontal,
-    }
-}
-
-fn browser_trail_config(
-    root_id: NodeId,
-    depth: u32,
-    direction: TrailDirection,
-    story: bool,
-) -> TrailConfigDto {
-    TrailConfigDto {
-        root_id,
-        mode: TrailMode::Neighborhood,
-        target_id: None,
-        depth,
-        direction,
-        caller_scope: TrailCallerScope::ProductionOnly,
-        edge_filter: Vec::new(),
-        show_utility_calls: false,
-        hide_speculative: false,
-        story,
-        node_filter: Vec::new(),
-        max_nodes: BROWSER_TRAIL_MAX_NODES,
-        layout_direction: LayoutDirection::Horizontal,
-    }
-}
-
-fn browser_trail_depth(value: Option<&str>) -> u32 {
-    value
-        .and_then(|value| value.parse::<u32>().ok())
-        .map(|value| value.min(BROWSER_TRAIL_MAX_DEPTH))
-        .unwrap_or(BROWSER_TRAIL_DEFAULT_DEPTH)
-}
-
-fn browser_trail_direction(value: Option<&str>) -> TrailDirection {
-    match value {
-        Some("incoming") => TrailDirection::Incoming,
-        Some("outgoing") => TrailDirection::Outgoing,
-        _ => TrailDirection::Both,
-    }
-}
-
-fn browser_bool_param(value: Option<&str>) -> bool {
-    matches!(
-        value.map(|value| value.to_ascii_lowercase()).as_deref(),
-        Some("1" | "true" | "yes" | "on")
-    )
-}
-
-fn browser_symbols_limit(value: Option<&str>) -> Option<u32> {
-    Some(
-        value
-            .and_then(|value| value.parse::<u32>().ok())
-            .map(|value| value.clamp(1, BROWSER_SYMBOLS_MAX_LIMIT))
-            .unwrap_or(BROWSER_SYMBOLS_DEFAULT_LIMIT),
-    )
-}
-
-fn http_target_selection_or_error(
-    stream: &mut TcpStream,
-    params: &HashMap<String, String>,
-) -> Result<Option<args::TargetSelection>> {
-    match target_selection_from_params(params) {
-        Ok(selection) => Ok(Some(selection)),
-        Err(error) => {
-            write_http_error_json(stream, 400, "invalid_target", error.to_string())?;
-            Ok(None)
-        }
-    }
-}
-
-fn resolve_http_target(
-    stream: &mut TcpStream,
-    runtime: &RuntimeContext,
-    target: args::TargetSelection,
-    file_filter: Option<&str>,
-) -> Result<Option<runtime::ResolvedTarget>> {
-    match resolve_target(runtime, target, file_filter) {
-        Ok(target) => Ok(Some(target)),
-        Err(error) => {
-            if let Some(ambiguous) = error.downcast_ref::<AmbiguousTargetError>() {
-                let output = build_ambiguous_target_error_output(&runtime.project_root, ambiguous);
-                write_http_json(stream, 400, &output)?;
-            } else {
-                write_http_error_json(stream, 400, "target_resolution_failed", error.to_string())?;
-            }
-            Ok(None)
-        }
-    }
-}
-
-fn target_selection_from_params(params: &HashMap<String, String>) -> Result<args::TargetSelection> {
-    if let Some(id) = params.get("id").filter(|value| !value.trim().is_empty()) {
-        return Ok(args::TargetSelection::Id(NodeId(id.trim().to_string())));
-    }
-    let query = params.get("q").cloned().unwrap_or_default();
-    if query.trim().is_empty() {
-        bail!("Pass `q` or `id`.");
-    }
-    Ok(args::TargetSelection::Query {
-        query,
-        choose: query_choose_param(params)?,
-    })
-}
-
-fn query_choose_param(params: &HashMap<String, String>) -> Result<Option<usize>> {
-    params
-        .get("choose")
-        .map(|value| {
-            value.parse::<usize>().with_context(|| {
-                format!("Invalid `choose` value `{value}`; expected a positive integer.")
-            })
-        })
-        .transpose()
-}
-
-fn write_http_json<T: serde::Serialize>(
-    stream: &mut TcpStream,
-    status: u16,
-    value: &T,
-) -> Result<()> {
-    let body = serde_json::to_string_pretty(value)?;
-    let status_text = match status {
-        200 => "OK",
-        400 => "Bad Request",
-        404 => "Not Found",
-        405 => "Method Not Allowed",
-        _ => "OK",
-    };
-    write!(
-        stream,
-        "HTTP/1.1 {status} {status_text}\r\nContent-Type: application/json\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{}",
-        body.len(),
-        body
-    )?;
-    Ok(())
-}
-
-fn parse_query_string(query: &str) -> HashMap<String, String> {
-    query
-        .split('&')
-        .filter(|part| !part.is_empty())
-        .filter_map(|part| {
-            let (key, value) = part.split_once('=').unwrap_or((part, ""));
-            Some((url_decode(key)?, url_decode(value)?))
-        })
-        .collect()
-}
-
-fn search_repo_text_mode_param(value: &str) -> Option<SearchRepoTextMode> {
-    match value.trim().to_ascii_lowercase().as_str() {
-        "auto" => Some(SearchRepoTextMode::Auto),
-        "on" | "true" | "1" => Some(SearchRepoTextMode::On),
-        "off" | "false" | "0" => Some(SearchRepoTextMode::Off),
-        _ => None,
-    }
-}
-
-fn url_decode(value: &str) -> Option<String> {
-    let mut out = Vec::with_capacity(value.len());
-    let bytes = value.as_bytes();
-    let mut idx = 0usize;
-    while idx < bytes.len() {
-        match bytes[idx] {
-            b'+' => out.push(b' '),
-            b'%' if idx + 2 < bytes.len() => {
-                let hex = std::str::from_utf8(&bytes[idx + 1..idx + 3]).ok()?;
-                out.push(u8::from_str_radix(hex, 16).ok()?);
-                idx += 2;
-            }
-            byte => out.push(byte),
-        }
-        idx += 1;
-    }
-    String::from_utf8(out).ok()
-}
-
-fn run_stdio_server(runtime: RuntimeContext) -> Result<()> {
-    let stdin = std::io::stdin();
-    let mut stdout = std::io::stdout();
-    for line in stdin.lock().lines() {
-        let line = line?;
-        if line.trim().is_empty() {
-            continue;
-        }
-        if let Some(response) = handle_stdio_message(&runtime, &line) {
-            writeln!(stdout, "{}", serde_json::to_string(&response)?)?;
-            stdout.flush()?;
-        }
-    }
-    Ok(())
-}
-
-fn handle_stdio_message(runtime: &RuntimeContext, line: &str) -> Option<serde_json::Value> {
-    let request: serde_json::Value = match serde_json::from_str(line) {
-        Ok(value) => value,
-        Err(error) => {
-            return Some(stdio_jsonrpc_error(
-                serde_json::Value::Null,
-                -32700,
-                format!("Parse error: {error}"),
-            ));
-        }
-    };
-    if !request.is_object() {
-        return Some(stdio_jsonrpc_error(
-            serde_json::Value::Null,
-            -32600,
-            "Invalid request: expected JSON-RPC object",
-        ));
-    }
-    let Some(id) = request.get("id").cloned() else {
-        return None;
-    };
-    let Some(method) = request.get("method").and_then(|value| value.as_str()) else {
-        return Some(stdio_jsonrpc_error(
-            id,
-            -32600,
-            "Invalid request: missing method",
-        ));
-    };
-    let legacy_response = match method {
-        "initialize" => {
-            return Some(stdio_jsonrpc_success(
-                id,
-                stdio_initialize_result_json(&request),
-            ));
-        }
-        "tools/list" => stdio_tools_list_json(),
-        "resources/list" => stdio_resources_list_json(),
-        "resources/templates/list" => stdio_resource_templates_list_json(),
-        "prompts/list" => stdio_prompts_list_json(),
-        "prompts/get" => {
-            let Some(name) = request
-                .pointer("/params/name")
-                .and_then(|value| value.as_str())
-            else {
-                return Some(stdio_jsonrpc_error(
-                    id,
-                    -32602,
-                    "Invalid params: missing prompt name",
-                ));
-            };
-            match stdio_prompt_get_json(name) {
-                Ok(response) => response,
-                Err(error) => {
-                    return Some(stdio_jsonrpc_error(id, -32602, error.to_string()));
-                }
-            }
-        }
-        "resources/read" => {
-            let Some(uri) = request
-                .pointer("/params/uri")
-                .and_then(|value| value.as_str())
-            else {
-                return Some(stdio_jsonrpc_error(
-                    id,
-                    -32602,
-                    "Invalid params: missing resource uri",
-                ));
-            };
-            read_stdio_resource(runtime, uri)
-        }
-        "tools/call" => {
-            let Some(name) = request
-                .pointer("/params/name")
-                .and_then(|value| value.as_str())
-            else {
-                return Some(stdio_jsonrpc_error(
-                    id,
-                    -32602,
-                    "Invalid params: missing tool name",
-                ));
-            };
-            if !is_stdio_tool_name(name) {
-                return Some(stdio_jsonrpc_error(
-                    id,
-                    -32602,
-                    format!("Unknown tool: {name}"),
-                ));
-            }
-            if request
-                .pointer("/params/arguments")
-                .is_some_and(|value| !value.is_object() && !value.is_null())
-            {
-                return Some(stdio_jsonrpc_error(
-                    id,
-                    -32602,
-                    "Invalid params: tool arguments must be an object",
-                ));
-            }
-            return Some(stdio_jsonrpc_tool_call_from_legacy(
-                id,
-                handle_stdio_tool_call(runtime, &request),
-            ));
-        }
-        _ => {
-            return Some(stdio_jsonrpc_error(
-                id,
-                -32601,
-                format!("Method not found: {method}"),
-            ));
-        }
-    };
-    Some(stdio_jsonrpc_from_legacy(id, legacy_response))
-}
-
-fn stdio_jsonrpc_success(id: serde_json::Value, result: serde_json::Value) -> serde_json::Value {
-    serde_json::json!({
-        "jsonrpc": "2.0",
-        "id": id,
-        "result": result
-    })
-}
-
-fn stdio_jsonrpc_error(
-    id: serde_json::Value,
-    code: i32,
-    message: impl Into<String>,
-) -> serde_json::Value {
-    serde_json::json!({
-        "jsonrpc": "2.0",
-        "id": id,
-        "error": {
-            "code": code,
-            "message": message.into()
-        }
-    })
-}
-
-fn stdio_jsonrpc_from_legacy(
-    id: serde_json::Value,
-    response: serde_json::Value,
-) -> serde_json::Value {
-    if let Some(result) = response.get("result") {
-        return stdio_jsonrpc_success(id, result.clone());
-    }
-    if let Some(error) = response.get("error") {
-        let message = stdio_legacy_error_message(error);
-        let code = if message.contains("unknown resource") {
-            -32602
-        } else {
-            -32000
-        };
-        let mut response = stdio_jsonrpc_error(id, code, message);
-        if error.is_object()
-            && let Some(error_object) = response.get_mut("error")
-            && let Some(error_object) = error_object.as_object_mut()
-        {
-            error_object.insert("data".to_string(), error.clone());
-        }
-        return response;
-    }
-    stdio_jsonrpc_success(id, response)
-}
-
-fn stdio_jsonrpc_tool_call_from_legacy(
-    id: serde_json::Value,
-    response: serde_json::Value,
-) -> serde_json::Value {
-    if let Some(result) = response.get("result") {
-        return stdio_jsonrpc_success(id, stdio_tool_call_success(result.clone()));
-    }
-    if let Some(error) = response.get("error") {
-        return stdio_jsonrpc_success(id, stdio_tool_call_error(error));
-    }
-    stdio_jsonrpc_success(id, stdio_tool_call_success(response))
-}
-
-fn stdio_tool_call_success(structured_content: serde_json::Value) -> serde_json::Value {
-    serde_json::json!({
-        "content": [
-            {
-                "type": "text",
-                "text": stdio_json_text(&structured_content)
-            }
-        ],
-        "structuredContent": structured_content
-    })
-}
-
-fn stdio_tool_call_error(error: &serde_json::Value) -> serde_json::Value {
-    let message = stdio_legacy_error_message(error);
-    let structured_content = if error.is_object() {
-        error.clone()
-    } else {
-        serde_json::json!({ "message": message.clone() })
-    };
-    serde_json::json!({
-        "content": [
-            {
-                "type": "text",
-                "text": message
-            }
-        ],
-        "structuredContent": structured_content,
-        "isError": true
-    })
-}
-
-fn stdio_legacy_error_message(error: &serde_json::Value) -> String {
-    error
-        .as_str()
-        .map(str::to_string)
-        .or_else(|| {
-            error
-                .get("message")
-                .and_then(|value| value.as_str())
-                .map(str::to_string)
-        })
-        .unwrap_or_else(|| error.to_string())
-}
-
-fn stdio_json_text(value: &serde_json::Value) -> String {
-    serde_json::to_string(value).unwrap_or_else(|_| value.to_string())
-}
-
-fn stdio_initialize_result_json(request: &serde_json::Value) -> serde_json::Value {
-    let protocol_version = request
-        .pointer("/params/protocolVersion")
-        .and_then(|value| value.as_str())
-        .unwrap_or("2024-11-05");
-    serde_json::json!({
-        "protocolVersion": protocol_version,
-        "name": "codestory",
-        "version": "0.1.0",
-        "serverInfo": {
-            "name": "codestory",
-            "version": "0.1.0"
-        },
-        "capabilities": {
-            "tools": {
-                "listChanged": false
-            },
-            "resources": {
-                "subscribe": false,
-                "listChanged": false
-            },
-            "prompts": {
-                "listChanged": false
-            }
-        }
-    })
-}
-
-fn handle_stdio_tool_call(
-    runtime: &RuntimeContext,
-    request: &serde_json::Value,
-) -> serde_json::Value {
-    let name = request
-        .pointer("/params/name")
-        .and_then(|value| value.as_str())
-        .unwrap_or_default();
-    let query = request
-        .pointer("/params/arguments/query")
-        .and_then(|value| value.as_str())
-        .unwrap_or_default()
-        .to_string();
-    match name {
-        "search" => handle_stdio_search(runtime, request, query),
-        "symbol" => handle_stdio_symbol(runtime, request),
-        "trail" => handle_stdio_trail(runtime, request),
-        "definition" => handle_stdio_definition(runtime, request),
-        "references" => handle_stdio_references(runtime, request),
-        "symbols" => handle_stdio_symbols(runtime, request),
-        "snippet" => handle_stdio_snippet(runtime, request),
-        "ask" => handle_stdio_ask(runtime, request, &query),
-        _ => serde_json::json!({"error": "unknown tool"}),
-    }
-}
-
-fn handle_stdio_search(
-    runtime: &RuntimeContext,
-    request: &serde_json::Value,
-    query: String,
-) -> serde_json::Value {
-    let repo_text = match request
-        .pointer("/params/arguments/repo_text")
-        .and_then(|value| value.as_str())
-    {
-        Some("on") => SearchRepoTextMode::On,
-        Some("off") => SearchRepoTextMode::Off,
-        _ => SearchRepoTextMode::Auto,
-    };
-    let limit_per_source = request
-        .pointer("/params/arguments/limit")
-        .and_then(|value| value.as_u64())
-        .map(|value| value.clamp(1, 50) as u32)
-        .unwrap_or(10);
-    runtime
-        .browser
-        .search_results(SearchRequest {
-            query,
-            repo_text,
-            limit_per_source,
-            hybrid_weights: None,
-            hybrid_limits: None,
-        })
-        .map(|result| serde_json::json!({"result": enrich_stdio_search_result(result)}))
-        .unwrap_or_else(|error| serde_json::json!({"error": map_api_error(error).to_string()}))
-}
-
-fn handle_stdio_symbol(runtime: &RuntimeContext, request: &serde_json::Value) -> serde_json::Value {
-    resolve_target(runtime, stdio_target_selection(request), None)
-        .and_then(|target| {
-            runtime
-                .browser
-                .symbol_context(target.selected.node_id)
-                .map_err(map_api_error)
-        })
-        .map(|result| serde_json::json!({"result": result}))
-        .unwrap_or_else(
-            |error| serde_json::json!({"error": stdio_legacy_error_value(runtime, &error)}),
-        )
-}
-
-fn handle_stdio_trail(runtime: &RuntimeContext, request: &serde_json::Value) -> serde_json::Value {
-    let direction = match request
-        .pointer("/params/arguments/direction")
-        .and_then(|value| value.as_str())
-    {
-        Some("incoming") => TrailDirection::Incoming,
-        Some("outgoing") => TrailDirection::Outgoing,
-        _ => TrailDirection::Both,
-    };
-    let depth = request
-        .pointer("/params/arguments/depth")
-        .and_then(|value| value.as_u64())
-        .map(|value| value.min(BROWSER_TRAIL_MAX_DEPTH as u64) as u32)
-        .unwrap_or(BROWSER_TRAIL_DEFAULT_DEPTH);
-    let story = request
-        .pointer("/params/arguments/story")
-        .and_then(|value| value.as_bool())
-        .unwrap_or(false);
-    resolve_target(runtime, stdio_target_selection(request), None)
-        .and_then(|target| {
-            runtime
-                .browser
-                .trail_context(browser_trail_config(
-                    target.selected.node_id,
-                    depth,
-                    direction,
-                    story,
-                ))
-                .map_err(map_api_error)
-        })
-        .map(|result| serde_json::json!({"result": result}))
-        .unwrap_or_else(
-            |error| serde_json::json!({"error": stdio_legacy_error_value(runtime, &error)}),
-        )
-}
-
-fn handle_stdio_definition(
-    runtime: &RuntimeContext,
-    request: &serde_json::Value,
-) -> serde_json::Value {
-    resolve_target(runtime, stdio_target_selection(request), None)
-        .and_then(|target| {
-            runtime
-                .browser
-                .definition_context(target.selected.node_id.clone())
-                .map_err(map_api_error)
-                .map(|symbol| {
-                    let node_id = target.selected.node_id.0.clone();
-                    let links = stdio_node_links(&node_id);
-                    let mut definition = serde_json::to_value(build_search_hit_output(
-                        &runtime.project_root,
-                        &target.selected,
-                        false,
-                    ))
-                    .unwrap_or_else(|error| serde_json::json!({"error": error.to_string()}));
-                    add_stdio_links(&mut definition, links.clone());
-                    serde_json::json!({
-                        "resolution": build_query_resolution_output(&runtime.project_root, &target),
-                        "definition": definition,
-                        "symbol": symbol,
-                        "links": links,
-                    })
-                })
-        })
-        .map(|result| serde_json::json!({"result": result}))
-        .unwrap_or_else(
-            |error| serde_json::json!({"error": stdio_legacy_error_value(runtime, &error)}),
-        )
-}
-
-fn handle_stdio_references(
-    runtime: &RuntimeContext,
-    request: &serde_json::Value,
-) -> serde_json::Value {
-    resolve_target(runtime, stdio_target_selection(request), None)
-        .and_then(|target| {
-            runtime
-                .browser
-                .references_context(browser_references_config(target.selected.node_id.clone()))
-                .map_err(map_api_error)
-        })
-        .map(|result| serde_json::json!({"result": result}))
-        .unwrap_or_else(
-            |error| serde_json::json!({"error": stdio_legacy_error_value(runtime, &error)}),
-        )
-}
-
-fn handle_stdio_symbols(
-    runtime: &RuntimeContext,
-    request: &serde_json::Value,
-) -> serde_json::Value {
-    let limit = request
-        .pointer("/params/arguments/limit")
-        .and_then(|value| value.as_u64())
-        .map(|value| value.clamp(1, BROWSER_SYMBOLS_MAX_LIMIT as u64) as u32)
-        .or(Some(BROWSER_SYMBOLS_DEFAULT_LIMIT));
-    let parent_id = request
-        .pointer("/params/arguments/parent_id")
-        .and_then(|value| value.as_str())
-        .filter(|value| !value.is_empty());
-    let result = if let Some(parent_id) = parent_id {
-        runtime
-            .browser
-            .list_children_symbols(ListChildrenSymbolsRequest {
-                parent_id: NodeId(parent_id.to_string()),
-            })
-            .map(|symbols| {
-                serde_json::to_value(symbols)
-                    .unwrap_or_else(|error| serde_json::json!({"error": error.to_string()}))
-            })
-    } else {
-        runtime
-            .browser
-            .list_root_symbols(ListRootSymbolsRequest { limit })
-            .map(|symbols| {
-                serde_json::to_value(symbols)
-                    .unwrap_or_else(|error| serde_json::json!({"error": error.to_string()}))
-            })
-    };
-    result
-        .map(|value| serde_json::json!({"result": {"symbols": value}}))
-        .unwrap_or_else(|error| serde_json::json!({"error": map_api_error(error).to_string()}))
-}
-
-fn handle_stdio_snippet(
-    runtime: &RuntimeContext,
-    request: &serde_json::Value,
-) -> serde_json::Value {
-    resolve_target(runtime, stdio_target_selection(request), None)
-        .and_then(|target| {
-            runtime
-                .browser
-                .snippet_context(target.selected.node_id, 4)
-                .map_err(map_api_error)
-        })
-        .map(|result| serde_json::json!({"result": result}))
-        .unwrap_or_else(
-            |error| serde_json::json!({"error": stdio_legacy_error_value(runtime, &error)}),
-        )
-}
-
-fn handle_stdio_ask(
-    runtime: &RuntimeContext,
-    request: &serde_json::Value,
-    default_prompt: &str,
-) -> serde_json::Value {
-    let prompt = request
-        .pointer("/params/arguments/prompt")
-        .and_then(|value| value.as_str())
-        .unwrap_or(default_prompt)
-        .to_string();
-    let max_results = request
-        .pointer("/params/arguments/max_results")
-        .and_then(|value| value.as_u64())
-        .map(|value| value.clamp(1, 50) as u32)
-        .unwrap_or(8);
-    let include_evidence = request
-        .pointer("/params/arguments/include_evidence")
-        .and_then(|value| value.as_bool())
-        .unwrap_or(true);
-    let response_mode = match request
-        .pointer("/params/arguments/response_mode")
-        .and_then(|value| value.as_str())
-    {
-        Some("markdown") => AgentResponseModeDto::Markdown,
-        _ => AgentResponseModeDto::Structured,
-    };
-    runtime
-        .browser
-        .ask(AgentAskRequest {
-            prompt,
-            retrieval_profile: codestory_contracts::api::AgentRetrievalProfileSelectionDto::Auto,
-            focus_node_id: None,
-            max_results: Some(max_results),
-            response_mode,
-            latency_budget_ms: None,
-            include_evidence,
-            hybrid_weights: None,
-            connection: AgentConnectionSettingsDto::default(),
-            run_local_agent: false,
-        })
-        .map(|result| serde_json::json!({"result": result}))
-        .unwrap_or_else(|error| serde_json::json!({"error": map_api_error(error).to_string()}))
-}
-
-fn stdio_target_selection(request: &serde_json::Value) -> args::TargetSelection {
-    if let Some(id) = request
-        .pointer("/params/arguments/id")
-        .and_then(|value| value.as_str())
-        .filter(|value| !value.trim().is_empty())
-    {
-        return args::TargetSelection::Id(NodeId(id.trim().to_string()));
-    }
-    let query = request
-        .pointer("/params/arguments/query")
-        .and_then(|value| value.as_str())
-        .unwrap_or_default()
-        .to_string();
-    args::TargetSelection::Query {
-        query,
-        choose: request
-            .pointer("/params/arguments/choose")
-            .and_then(|value| value.as_u64())
-            .map(|value| value as usize),
-    }
-}
-
-fn stdio_legacy_error_value(runtime: &RuntimeContext, error: &anyhow::Error) -> serde_json::Value {
-    if let Some(ambiguous) = error.downcast_ref::<AmbiguousTargetError>() {
-        return serde_json::to_value(build_ambiguous_target_error_output(
-            &runtime.project_root,
-            ambiguous,
-        ))
-        .ok()
-        .and_then(|value| value.get("error").cloned())
-        .unwrap_or_else(|| serde_json::json!(ambiguous.to_string()));
-    }
-
-    serde_json::json!(error.to_string())
-}
-
-fn read_stdio_resource(runtime: &RuntimeContext, uri: &str) -> serde_json::Value {
-    let result = match uri {
-        "codestory://status" => read_stdio_status_resource(runtime),
-        "codestory://agent-guide" => Ok(read_stdio_agent_guide_resource()),
-        "codestory://project" => runtime
-            .open_project_summary()
-            .map(|summary| serde_json::json!(summary)),
-        "codestory://grounding" => runtime
-            .grounding
-            .grounding_snapshot(GroundingBudgetDto::Balanced)
-            .map(|snapshot| serde_json::json!(snapshot))
-            .map_err(map_api_error),
-        "codestory://symbols/root" => runtime
-            .browser
-            .list_root_symbols(ListRootSymbolsRequest {
-                limit: Some(BROWSER_SYMBOLS_DEFAULT_LIMIT),
-            })
-            .map(|symbols| serde_json::json!(symbols))
-            .map_err(map_api_error),
-        _ => read_stdio_template_resource(runtime, uri),
-    };
-    result
-        .map(|value| serde_json::json!({"result": {"contents": [{"uri": uri, "mimeType": "application/json", "text": value.to_string()}]}}))
-        .unwrap_or_else(|error| serde_json::json!({"error": error.to_string()}))
-}
-
-fn read_stdio_status_resource(runtime: &RuntimeContext) -> Result<serde_json::Value> {
-    let summary = runtime.open_project_summary()?;
-    let retrieval = summary.retrieval.as_ref();
-    Ok(serde_json::json!({
-        "project_root": crate::display::clean_path_string(&runtime.project_root.to_string_lossy()),
-        "storage_path": crate::display::clean_path_string(&runtime.storage_path.to_string_lossy()),
-        "storage_exists": runtime.storage_path.exists(),
-        "retrieval_mode": retrieval.map(|state| state.mode),
-        "semantic_ready": retrieval.is_some_and(|state| state.semantic_ready),
-        "semantic_doc_count": retrieval.map(|state| state.semantic_doc_count).unwrap_or(0),
-        "fallback_reason": retrieval.and_then(|state| state.fallback_reason),
-        "fallback_message": retrieval.and_then(|state| state.fallback_message.as_deref()),
-        "index_freshness": summary.freshness,
-        "recommended_next_calls": [
-            {
-                "method": "resources/read",
-                "uri": "codestory://agent-guide"
-            },
-            {
-                "method": "tools/call",
-                "tool": "search",
-                "arguments": {
-                    "query": "<symbol-or-concept>",
-                    "limit": 10
-                }
-            },
-            {
-                "method": "tools/call",
-                "tool": "definition",
-                "arguments": {
-                    "id": "<node_id-from-search>"
-                }
-            },
-            {
-                "method": "resources/read",
-                "uri": "codestory://trail/<node_id-from-search>"
-            }
-        ]
-    }))
-}
-
-fn read_stdio_agent_guide_resource() -> serde_json::Value {
-    serde_json::json!({
-        "purpose": "Default read-only CodeStory browser loop for local codebase grounding.",
-        "recommended_call_sequence": [
-            {
-                "method": "resources/read",
-                "uri": "codestory://status"
-            },
-            {
-                "method": "tools/call",
-                "tool": "search",
-                "arguments": {
-                    "query": "<symbol-or-task>",
-                    "limit": 10
-                }
-            },
-            {
-                "method": "tools/call",
-                "tool": "definition",
-                "arguments": {
-                    "id": "<best-node-id>"
-                }
-            },
-            {
-                "method": "resources/read",
-                "uri": "codestory://snippet/<best-node-id>"
-            },
-            {
-                "method": "resources/read",
-                "uri": "codestory://references/<best-node-id>"
-            },
-            {
-                "method": "resources/read",
-                "uri": "codestory://trail/<best-node-id>"
-            }
-        ],
-        "safety_notes": [
-            "All stdio tools are read-only, non-destructive, idempotent, local-only, and closed-world.",
-            "Use continuation links from search or definition results before broadening retrieval.",
-            "Keep search limits bounded; stdio search clamps limit to 1..50.",
-            "Treat repo-text hits as evidence until a resolvable symbol is selected."
-        ]
-    })
-}
-
-fn enrich_stdio_search_result(
-    result: codestory_contracts::api::SearchResultsDto,
-) -> serde_json::Value {
-    let mut value = serde_json::to_value(result)
-        .unwrap_or_else(|error| serde_json::json!({"error": error.to_string()}));
-    for field in [
-        "suggestions",
-        "indexed_symbol_hits",
-        "repo_text_hits",
-        "hits",
-    ] {
-        if let Some(hits) = value.get_mut(field).and_then(|field| field.as_array_mut()) {
-            for hit in hits {
-                enrich_stdio_search_hit(hit);
-            }
-        }
-    }
-    value
-}
-
-fn enrich_stdio_search_hit(hit: &mut serde_json::Value) {
-    if !hit
-        .get("resolvable")
-        .and_then(|value| value.as_bool())
-        .unwrap_or(false)
-    {
-        return;
-    }
-    let Some(node_id) = hit
-        .get("node_id")
-        .and_then(|value| value.as_str())
-        .filter(|value| !value.is_empty())
-        .map(str::to_string)
-    else {
-        return;
-    };
-    add_stdio_links(hit, stdio_node_links(&node_id));
-}
-
-fn add_stdio_links(hit: &mut serde_json::Value, links: serde_json::Value) {
-    if let Some(object) = hit.as_object_mut() {
-        object.insert("links".to_string(), links);
-    }
-}
-
-fn stdio_node_links(node_id: &str) -> serde_json::Value {
-    serde_json::json!([
-        {
-            "rel": "symbol",
-            "uri": format!("codestory://symbol/{node_id}")
-        },
-        {
-            "rel": "snippet",
-            "uri": format!("codestory://snippet/{node_id}")
-        },
-        {
-            "rel": "references",
-            "uri": format!("codestory://references/{node_id}")
-        },
-        {
-            "rel": "trail",
-            "uri": format!("codestory://trail/{node_id}")
-        }
-    ])
-}
-
-fn read_stdio_template_resource(runtime: &RuntimeContext, uri: &str) -> Result<serde_json::Value> {
-    let Some((kind, node_id)) = uri
-        .strip_prefix("codestory://")
-        .and_then(|tail| tail.split_once('/'))
-    else {
-        bail!("unknown resource");
-    };
-    let node_id = NodeId(node_id.to_string());
-    match kind {
-        "symbol" => runtime
-            .browser
-            .symbol_context(node_id)
-            .map(|value| serde_json::json!(value))
-            .map_err(map_api_error),
-        "references" => runtime
-            .browser
-            .references_context(browser_references_config(node_id))
-            .map(|value| serde_json::json!(value))
-            .map_err(map_api_error),
-        "snippet" => runtime
-            .browser
-            .snippet_context(node_id, 4)
-            .map(|value| serde_json::json!(value))
-            .map_err(map_api_error),
-        "trail" => runtime
-            .browser
-            .trail_context(browser_trail_config(
-                node_id,
-                BROWSER_TRAIL_DEFAULT_DEPTH,
-                TrailDirection::Both,
-                false,
-            ))
-            .map(|value| serde_json::json!(value))
-            .map_err(map_api_error),
-        _ => bail!("unknown resource"),
-    }
-}
-
 fn ensure_dot_only_for_trail(format: args::OutputFormat, command: &str) -> Result<()> {
     if format == args::OutputFormat::Dot {
         bail!("--format dot is only supported by `trail`; `{command}` supports markdown and json");
@@ -3678,24 +1718,6 @@ fn is_speculative_certainty(certainty: Option<&str>) -> bool {
     matches!(
         certainty.map(|value| value.to_ascii_lowercase()).as_deref(),
         Some("uncertain" | "speculative")
-    )
-}
-
-fn write_http_error_json(
-    stream: &mut TcpStream,
-    status: u16,
-    code: &'static str,
-    message: impl Into<String>,
-) -> Result<()> {
-    write_http_json(
-        stream,
-        status,
-        &serde_json::json!({
-            "error": {
-                "code": code,
-                "message": message.into()
-            }
-        }),
     )
 }
