+++ crates/codestory-cli/tests/search_json_output.rs
@@ -21,12 +21,15 @@ impl EnvGuard {
 
 impl Drop for EnvGuard {
     fn drop(&mut self) {
-        unsafe {
-            if let Some(value) = self.previous.as_deref() {
-                std::env::set_var(self.key, value);
-            } else {
-                std::env::remove_var(self.key);
-            }
+        restore_env_value(self.key, self.previous.as_deref());
+    }
+}
+
+fn restore_env_value(key: &'static str, previous: Option<&str>) {
+    unsafe {
+        match previous {
+            Some(value) => std::env::set_var(key, value),
+            None => std::env::remove_var(key),
         }
     }
 }
