+++ crates/codestory-runtime/tests/retrieval_eval.rs
@@ -34,24 +34,19 @@ impl EnvGuard {
         }
         Self { key, previous }
     }
-
-    fn remove(key: &'static str) -> Self {
-        let previous = std::env::var(key).ok();
-        unsafe {
-            std::env::remove_var(key);
-        }
-        Self { key, previous }
-    }
 }
 
 impl Drop for EnvGuard {
     fn drop(&mut self) {
-        unsafe {
-            if let Some(value) = self.previous.as_deref() {
-                std::env::set_var(self.key, value);
-            } else {
-                std::env::remove_var(self.key);
-            }
+        restore_env_value(self.key, self.previous.as_deref());
+    }
+}
+
+fn restore_env_value(key: &'static str, previous: Option<&str>) {
+    unsafe {
+        match previous {
+            Some(value) => std::env::set_var(key, value),
+            None => std::env::remove_var(key),
         }
     }
 }
